reorganize files and update test suite for initial unit tests
