+++ client/src/components/Overview/Image Gallery/DefaultGallery.jsx
