+++ client/src/components/Overview/Image Gallery/ExpandedGallery.jsx
