+++ client/src/components/Overview/Product Information/Price.jsx
@@ -8,7 +8,7 @@ const Price = (props) => {
   }
 
   return (
-    <div className="price">
+    <div>
       <h4>Overview - Price [from Style]</h4>
       {defaultPrice}
     </div>
