+++ client/src/components/Overview/Style Selector/__test__/OverviewStyleSelector.test.js
@@ -1,0 +1,22 @@
+/**
+ * @jest-environment jsdom
+ */
+
+import { fireEvent, render, screen } from '@testing-library/react'
+import { queries } from '@testing-library/dom'
+import userEvent from '@testing-library/user-event'
+import '@testing-library/jest-dom'
+import React, { useEffect, useState } from "react"
+
+// component
+import { Overview } from '../../../Overview.jsx'
+// subcomponents
+import StyleSelector from '../StyleSelector.jsx'
+
+test('sample', () => {
+
+})
+
+// IDEAS
+// test for props
+// test for state
